Fix pip executing on every run
The version check used `is not`operator. Since `is not` checks  if the
instances are same and does not compare strings, pip was always invoked
whenever `gauge run` is executed.

This patch changes `is not` to `!=` which correctly compares strings. A
check is added to properly report error when plugin is not installed.

The `subprocess.check_output`  function is used as it will detect
subprocess failures and raise exception.